requirements.txt
fastapi
uvicorn
playwright
pillow